+++ mitra_activitypub/src/handlers/activity.rs
@@ -4,12 +4,21 @@ use serde_json::{Value as JsonValue};
 
 use apx_sdk::deserialization::object_to_id;
 use mitra_config::Config;
-use mitra_models::database::DatabaseClient;
+use mitra_models::{
+    activitypub::queries::{
+        add_object_to_collection,
+        save_activity,
+    },
+    database::{DatabaseClient, DatabaseError},
+    users::queries::get_portable_user_by_actor_id,
+};
 use mitra_validators::errors::ValidationError;
 
 use crate::{
-    forwarder::get_activity_audience,
-    ownership::verify_activity_owner,
+    forwarder::{get_activity_audience, get_activity_recipients},
+    identifiers::canonicalize_id,
+    ownership::{get_object_id, verify_activity_owner},
+    queues::OutgoingActivityJobData,
     vocabulary::*,
 };
 
@@ -63,14 +72,18 @@ pub async fn handle_activity(
 ) -> Result<(), HandlerError> {
     // Validate common activity attributes
     verify_activity_owner(activity)?;
+    let activity_id = get_object_id(activity)?;
+    let canonical_activity_id = canonicalize_id(activity_id)?;
     let activity_type = activity["type"].as_str()
         .ok_or(ValidationError("type property is missing"))?
         .to_owned();
     let activity_actor = object_to_id(&activity["actor"])
         .map_err(|_| ValidationError("invalid actor property"))?;
-    let _audience = get_activity_audience(activity)?;
+    let canonical_actor_id = canonicalize_id(&activity_actor)?;
+    let audience = get_activity_audience(activity)?;
 
     let activity = activity.clone();
+    let activity_clone = activity.clone();
     let maybe_descriptor = match activity_type.as_str() {
         ACCEPT => {
             handle_accept(config, db_client, activity).await?
@@ -126,6 +139,59 @@ pub async fn handle_activity(
         },
     };
     if let Some(descriptor) = maybe_descriptor {
+        save_activity(
+            db_client,
+            &canonical_activity_id.to_string(),
+            &activity_clone,
+        ).await?;
+        let recipients = get_activity_recipients(
+            db_client,
+            &config.instance_url(),
+            &audience,
+        ).await?;
+        for recipient in recipients.iter() {
+            // Recipient is a local actor: add activity to its inbox
+            if recipient.has_account() && recipient.is_portable() {
+                add_object_to_collection(
+                    db_client,
+                    recipient.id,
+                    &recipient.expect_actor_data().inbox,
+                    &canonical_activity_id.to_string(),
+                ).await?;
+            };
+        };
+        match get_portable_user_by_actor_id(
+            db_client,
+            &canonical_actor_id.to_string(),
+        ).await {
+            Ok(actor) => {
+                // Activity has been performed by a local actor:
+                // add to outbox and forward
+                add_object_to_collection(
+                    db_client,
+                    actor.id,
+                    &actor.profile.expect_actor_data().outbox,
+                    &canonical_activity_id.to_string(),
+                ).await?;
+                let remote_recipients = recipients.iter()
+                    .filter_map(|recipient| recipient.actor_json.clone())
+                    .collect();
+                // Forward only if HTTP signature can be created
+                if let Some(job_data) = OutgoingActivityJobData::new_forwarded(
+                    &config.instance_url(),
+                    &actor,
+                    &activity_clone,
+                    remote_recipients,
+                ) {
+                    // Activity has already been saved
+                    job_data.enqueue(db_client).await?;
+                } else {
+                    log::warn!("signing keys are not found in actor document");
+                };
+            },
+            Err(DatabaseError::NotFound(_)) => (),
+            Err(other_error) => return Err(other_error.into()),
+        };
         log::info!(
             "processed {}({}) from {}",
             activity_type,